cron: fix type in comparison
Use of size_t in this context should be also safer from bit shift
perspective, risk of overflowing it is smaller.

cronnext.c:80:16: warning: comparison between signed and unsigned integer
expressions [-Wsign-compare]
  for (f = 1; f < sizeof(flagname);  f = f << 1)
                ^

+++ src/cronnext.c
@@ -81,7 +81,7 @@ const char *flagname[]= {
 };
 
 void printflags(char *indent, int flags) {
-	int f;
+	size_t f;
 	int first = 1;
 
 	printf("%s    flagnames:", indent);
